+++ sdk/storage/azure-storage-file-datalake/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 12.4.0-beta.1 (Unreleased)
 - Added support to list paths on a directory.
+- Exposed ClientOptions on all client builders, allowing users to set a custom application id and custom headers.
 - Fixed a bug where the error message would not be displayed the exception message of a HEAD request.
 
 ## 12.3.0 (2020-11-11)